JCS5300, JCS5400 加入檢核。

+++ tw.com.citi.catalog.web/src/tw/com/citi/catalog/web/pages/JCS5400.java
@@ -4,9 +4,12 @@ import java.util.Map;
 
 import org.apache.wicket.PageParameters;
 import org.apache.wicket.spring.injection.annot.SpringBean;
+import org.springframework.util.StringUtils;
 
 import tw.com.citi.catalog.web.dao.ICoordinatorDao;
+import tw.com.citi.catalog.web.dao.IScrDao;
 import tw.com.citi.catalog.web.grid.IGridHandler;
+import tw.com.citi.catalog.web.model.Coordinator;
 import tw.com.citi.catalog.web.utils.AccessControlUtil;
 
 import com.google.gson.Gson;
@@ -19,6 +22,9 @@ public class JCS5400 extends AbstractBasePage {
 
     @SpringBean(name = "coordinatorDao")
     private ICoordinatorDao coordinatorDao;
+
+    @SpringBean(name = "scrDao")
+    private IScrDao scrDao;
 
     private transient Gson gson = new Gson();
 
@@ -34,19 +40,39 @@ public class JCS5400 extends AbstractBasePage {
     public String handleRequest(PageParameters params) {
         String actionName = params.getString("actionName");
         String actionParams = params.getString("actionParams");
-        Map dataMap = gson.fromJson(actionParams, new TypeToken<Map<String, String>>() {}.getType());
-        if(AccessControlUtil.authenticateCBCUser(dataMap)) {
+        Map dataMap = gson.fromJson(actionParams, new TypeToken<Map<String, String>>() {
+        }.getType());
+        String name = (String) dataMap.get("name");
+        String team = (String) dataMap.get("team");
+        Coordinator coordinator = coordinatorDao.findUnique(name, team);
+        if (AccessControlUtil.authenticateCBCUser(dataMap) && doCheck(name, team)) {
             if ("Create".equals(actionName)) {
+                if (coordinator != null) {
+                    throw new IllegalStateException("Coordinator with given name and team is already exist.");
+                }
                 coordinatorDao.create(dataMap);
             } else if ("Modify".equals(actionName)) {
+                if (coordinator != null) {
+                    throw new IllegalStateException("Coordinator with given name and team is already exist.");
+                }
                 coordinatorDao.update(dataMap);
             } else if ("Delete".equals(actionName)) {
+                if (scrDao.countByProgrammerId(coordinator.getId()) > 0) {
+                    throw new IllegalStateException("Cannot delete Coordinator that related to exist SCR.");
+                }
                 coordinatorDao.delete(dataMap);
             }
         } else {
             throw new IllegalArgumentException("Authentication fail");
         }
         return super.handleRequest(params);
+    }
+
+    private boolean doCheck(String name, String team) {
+        if (!StringUtils.hasText(name) || !StringUtils.hasText(team)) {
+            throw new IllegalArgumentException("Cannot find Programmer Name or Team.");
+        }
+        return true;
     }
 
 }
